feat: rename extension folder
