+++ lib/modules/chat/domain/usecases/get_message_histories.dart
@@ -1,4 +1,4 @@
-import 'package:flutter_assessment_test/core/extension/future_ext.dart';
+import 'package:flutter_assessment_test/core/extensions/future_ext.dart';
 import 'package:flutter_assessment_test/core/types/resource.dart';
 import 'package:flutter_assessment_test/core/types/usecase.dart';
 import 'package:flutter_assessment_test/modules/chat/domain/entities/chat_message_histories.dart';
